US91774 Delay loading of image until tile is visible
This uses IntersectionObserver to only load the image when it's actually within the user's viewport. This means we can defer loading of images until we strictly require them.

Also did some cleanup and re-jiggering to make use of computed properties and observers a bit more (which were tangentially required so that we can control when the img element's `src` and `srcset` attributes were set anyway).